fix:add nodes
